backend/requirements.txt
gunicorn==20.1.0
Django==3.2.16
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==9.3.0
PyJWT==2.1.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
requests==2.26.0
django-filter==21.1
cryptography==41.0.2
djoser
PyYAML==6.0
drf-extra-fields
drf-yasg
